pyproject.toml

[tool.poetry]
name = "cdf_ts_replicator"
version = "1.0.0"
description = "Stream time series from one CDF project to another using time series subscriptions"
authors = ["Murad Saeter <murad.sater@cognite.com>"]

[tool.black]
line-length = 120
target_version = ['py38']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["cognite"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
cognite-extractor-utils = "^5.0.0"
cognite-extractor-utils-rest = {path = "../python-extractor-utils-rest/dist/cognite_extractor_utils_rest-0.6.0-py3-none-any.whl"}

[tool.poetry.dev-dependencies]
pyinstaller = "^4.7"
macholib = {version = "^1.14", platform = "darwin"}             # Used by pyinstaller pn Mac OS
pywin32-ctypes = {version = "^0.2.0", platform = "win32"}       # Used by pyinstaller on Windows
pefile = "^2019.4.18"                                           # Used by pyinstaller on Windows

[tool.poetry.scripts]
cdf_ts_replicator = "cdf_ts_replicator.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.cogex.docker]
tags = ["cognite/my-extractor:{version}"]
